+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### 🐛 Bug fixes
 
+- Changed `x`, `y`, `width`, and `height` to refer to the top left of the outer box, and draw inner
+
 ### 📚 Documentation and demo website changes
 
 ### 🔧 Internal changes